LayoutTests/accessibility/mac/treeitem-row-delegation.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<script src="../../resources/accessibility-helper.js"></script>
</head>

<body id="body">

<div id="tree" role="tree">
    <div id="item-1" role="treeitem" aria-level="1" aria-expanded="true" tabindex="0">Jungle Animals
        <div role="group" aria-owns="group1-item1 group1-item2 group1-item3"></div>
    </div>
    <div id="group1-item1" role="treeitem" aria-level="2" aria-setsize="3" aria-posinset="1" aria-current="page">Lions</div>
</div>

<script>
    description("This tests that a treeitem was a group that uses aria-owns will report its disclosed rows correctly.");

    if (window.accessibilityController) {
       var treeitem = accessibilityController.accessibleElementById("item-1");
       var treeitem2 = treeitem.disclosedRowAtIndex(0);
       shouldBeTrue("treeitem2.isEqual(accessibilityController.accessibleElementById('group1-item1'))");
    }
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
